Replaced interface AttachmentsReader with implementation
